refactored feedback page to include a new feedback form component
